Reorganize usePressEvents and PressResponder
