+++ packages/react-native-web/src/hooks/usePressEvents/index.js
@@ -10,7 +10,9 @@
 
 'use strict';
 
-import PressResponder, { type EventHandlers, type PressResponderConfig } from './index';
+import type { EventHandlers, PressResponderConfig } from './PressResponder';
+
+import PressResponder from './PressResponder';
 import { useEffect, useRef } from 'react';
 
 export default function usePressEvents(hostRef: any, config: PressResponderConfig): EventHandlers {
